test: e2e tests remove user (see details)
this is complicated and took some research. Rux styling hides a lot of html elements. The work around is to go into the shadow content, give the actual html form element some height and width, then use cypress to force it to show. Now it can be interacted with by ignoring the Rux content and working with the shadow content.